Add appName and stratergies
